requirements.txt
google-generativeai 
python-dotenv
google-api-python-client